Rename files to match the convention of other services and edit the matching functions
Files will be renamed again in the next commit
